Include base version in nightly versioning scheme
Changes nightly versions from:
```
0.0.0-nightly.<date>+<commithash>
```
to:
```
0.0.0-nightly-<version>.<date>+<commithash>
```

This will make it easier to identify versions in the big list of
nightlies. The version number is appended in the prerelease
metadata part of the semver version, because nightlies actually do _not_
have to be in any way semver-compatible with that version (they can even
include major breaking changes).