Visibility tweaks for the connectors mapping playground.
We can deploy https://connectors-mapping.netlify.app from a local branch
if we have to, but it would be nice to have a stable apollo-federation
version that is fully usable by the playground, as-is.

The last time I investigated updating the playground to
`ConnectSpec::V0_3`, I found these changes to be necessary:
https://github.com/apollographql/connectors-mapping-playground/pull/76

Those changes in turn require making `HeaderValue::parse_with_spec` and
`SourceHTTPArguments::from_directive` public so they can be used by the
mapping playground (which is what this commit does).