Allow `null` as successful default in `$(missing ?? null)`.
Since the final operand does not have a `??` operator after it, I
believe a final `null` value should not be mapped to `None`, but instead
should be returned as-is, as a successful chain result.

At a more intuitive level, if `$(possibly_missing ?? any_default)`
becomes our standard idiom for default values, we shouldn't have to
explain/maintain the exception that `any_default` can't be `null` when
you're using a chain of `??` operators (but can be `null` when you're
using `?!`). With this commit, any `any_default` result counts as
successful as long as it's not `None`.