Use `SchemaTypeRef` instead of `Node<ObjectType>` in `ConnectedElement`.
A SchemaTypeRef is a reference to a &Schema as well as the &Name and
&ExtendedType of some named type within the schema, all with the same
'schema lifetime. The SchemaTypeRef::new method returns an
Option<SchemaTypeRef> to guarantee that SchemaTypeRef references are
only created for actual schema types.

Because SchemaTypeRef has access to the Schema (not just an individual
element), it can perform operations that are aware of other types within
the same schema, like finding all the implementing/member types of an
interface or union.

Because SchemaTypeRef can represent any named type in the schema, not
just ObjectType, these changes should lay some of the necessary
groundwork for working with abstract types, where the type in question
might be an InterfaceType or UnionType.

After some earlier attempts, I've decided it's important to use a type
like SchemaTypeRef that still implements the Copy trait, rather than
(say) ExtendedType, so we don't have to update all the data structures
that assume they can implement Copy while using ConnectedElement.