fix(federation/composition): Don't error for `@join__field` on an input object field during API schema query graph creation (#8132)
When creating the API schema query graph during composition, we collect which fields have progressive overrides along with their labels. We extract this from `@join__field` applications in the supergraph schema, but this extraction code was mistakenly bailing when `@join__field` was seen on an input object field (which they're allowed to be on). We now instead ignore such applications during extraction, as input object fields cannot be overridden.