calculate cache keys and policies for connector requests
cache keys are lazily-computed copies of HTTP details. cache policies are cache-control headers. in fetch nodes that call root fields, the keys and policies apply to the entire fetch node (if there are more than one, then the operation uses aliases. but we can't split that fetch node up further, so cache keys should be combined and cache policies should use the most restrictive.) in fetch nodes for entities, there's one key/policy per entity in the `representations` variable, in the same order. this allows entities to be cached independently.