fix: pipeline shutdown race condition fails to shut down connections
Previously we were using Notify to indicate to connections that they should shut down when a reload happens. However, this creates a race condition where if a connection is being established at the moment that shutdown is called the new connection will never recieve the shutdown notification.

This PR switches to CancellationToken which once cancelled will always yield cancelled immediatey.