Update usage of `shape` crate for 0.7.0-preview.n releases.
We are updating the `shape` crate before introducing any abstract
types-related changes, because this update benefits all connect spec
versions, while still allowing us to preserve backwards/forwards
compatibility for older versions (connect/v0.3 and earlier). It would be
a shame to have to maintain two slightly different `shape` APIs (hogging
the same cargo package name) in order to keep connect/v0.3 and earlier
working as before.

One of the goals of this commit is to show how limited/reasonable the
updates are, because the new functionality coming in shape@0.7.0 does
not radically alter the behavior of the library (at least not in the
ways we were previously using it).

Some notable (expected) changes: we've started using a
`shape.locations()` method instead of accessing/mutating the
`shape.locations` vector directly, the new shape name tracking system
allows better error messages, and shape pretty-printing may now include
newlines for readability.