fix: release builds of docker images should utilize correct target platform
When making #8056, I missed an important detail in my analysis of the
CircleCI script responsible for building the router releases.  In
particular, I failed to observe that the `--platform` argument was variable
and could be both arm64 and amd64 varieties and actually requests [BOTH],
which packages two within one container [multi-build style].

This change updates the `download_and_validate.sh` script to make use of a
Docker provided `TARGETPLATFORM` variable in deciding which release tarball
to place within the built image, depending on the platform.

[BOTH]: https://github.com/apollographql/router/blob/0f801b8703bdc4abe7f56b4c4b3a94bfad518fec/.circleci/config.yml#L924-L927
[multi-build style]: https://docs.docker.com/build/building/multi-platform/