fix: don't attempt to carryover directives for missing types
if a type is unused (no connector references it) it won't end up in the supergraph. but if it has a directive we usually need to preserve, the code would try to add that directive. this change ignores missing types, which weren't reachable anyway to it has no effect on planning