docs: tweak to note on hot-reload
Clarified behavior around hot-reload based on customer feedback. I realize there's more to add here, but this addresses the confusion about how it works. (Even if it's not necessarily the most recommended way and has other non-deterministic concerns, it doesn't inherently cause downtime in the way in which it's implemented, which was the point of confusion.