Fix regression: skip executing SubSelection when input_shape None.
In PR #8221 I pushed down the blanket check for `input_shape.is_none()`
at the beginning of `PathList::compute_output_shape`, so it now occurs
only for variants that require access to the input shape, like
`PathList::Key` and `PathList::Method`, while also allowing `$variable`
evaluation and the `$(...)` syntax even when given a `None` input shape,
since those PathList variants do not necessarily have to use the given
input shape.

Unfortunately, I missed a case where the `None`-termination is still
important: we do not want to execute the `SubSelection` of any `Path
SubSelection` if the final shape of the `Path` part evaluates to None.
That was the previous behavior, and this commit restores it, with some
additional regression tests.