Refactor subgraph-side subscription special-casing
Now it's all encapsulated in this `subgraph_request` function inside the
subscription plugin.

By checking that it's actually a subscription request as the very first
thing, we can avoid some `if`/`else` conditions and `Option` handling
in the refactored version.

This code now has a shape that is very suitable to be pulled into a
tower layer!