Disable parsing `...` spread syntax until connect/v0.4.
What really happens: we still optionally parse the `...` token, but if
you use it (and `spec` is not magically `> ConnectSpec::V0_3`), you get
a fatal parse error explaining that `...` syntax is planned for
connect/v0.4. The fatality of this error depends on the previous commit,
which made `span.extra.errors` more consequential.

Note that we still support the kind of spread syntax where the path must
have a trailing `SubSelection` and no leading `...` token, so we still
need most of the machinery for handling spreads, but `...` enables
union-of-objects output types, which we are not fully prepared to
represent in validation or expansion (at least not in connect/v0.3).