Use `local_var_names` to suggest typo'd local variables.
https://github.com/apollographql/router/pull/8259#discussion_r2379515704

To support this analysis, I refactored the existing `ExternalVarPaths`
trait to be named `VarPaths` and require implementing a `var_paths`
method that returns a `Vec<&PathSelection>` of all external and local
variable references, which is now consumed by both
`VarPaths::external_var_paths` (same as the old `external_var_paths`
method) and the new `VarPaths::local_var_names` method.

This analysis is not completely precise about which local variables are
actually available at the point of reference, but that would require
more metadata about the JSONSelection than we currently have when the
relevant errors are generated during validation.