docs: extensive editorial and stylization
Built on the incredible foundation that Benjamin started, this does a lot of content editorial that is strongly informed by the style guide.  It consolidates 10 pages down to 6 focused pages by merging related content and using a different information architecture.

Page merges:
- TTL + Invalidation ->Cache Expiration
- Configuration + Storage + Private Data -> Customization
- Telemetry + Debugging -> Observability

Also, fixed frontmatter, and reorganized FAQ.

Honestly, a LOT of changes, but I think it matches our docs pretty nicely at
this point.  I could certainly use a technical review again though, and it's
important to push back on anything not on the mark or that is confusing.