ci: allow websocket test flakes
These are all flakey right now due to timing races when the websocket
closes. The failures are benign, it's just that the assertion sometimes
happens before the websocket has closed.

I'm not sure how to fix this right now, adding a bunch of artificial
delays doesn't seem ideal, so it seems better to just retry them for the
time being.

(One possible direction could be to refactor the tests so they use a
channel to signal close, and then tests can wait for that signal with a
timeout.)