fix: update auth plugin handling of directive renames
The router auth plugin is not properly handling when subgraphs rename their auth directives via imports. When such renames are made, the `@link`-processing code in the plugin ignores imports completely, leading to auth constraints imposed by renamed directives being ignored.

This PR adds tests to illustrate the issue and then fixes it by updating plugin logic to call the appropriate code in the `apollo-federation` crate (which handles both `as` and `imports` properly).