fix(federation/query-planner): Avoid unnecessarily precomputing non-trivial followup edges (#8373)
We currently precompute non-trivial followup edge information for all query graph builders, so as to mirror the JS codebase. This is necessary when building federated query graphs and supergraph/API schema query graphs. However, this isn't needed when building subgraph query graphs (as `FederatedQueryGraphBuilder::build()` does it later, and no graph paths are used in between).

This PR accordingly updates `SchemaQueryGraphBuilder::build()` to skip this precomputation if building a subgraph query graph, to avoid the unnecessary edge iteration/memory allocation. Note that this precomputation was originally skipped completely in `SchemaQueryGraphBuilder::build()` during query planning, but was introduced when aligning parity with the JS codebase in https://github.com/apollographql/router/pull/7961.