fix(query-planner,composition): Remove unused overriding path IDs (#8502)
We created overriding path IDs back in the JS query planner to support sorting closed branch options better, leading to more efficient query plans when we had to trim options. But this sorting would sometimes cause QP to be very expensive, so we ditched it. We didn't stop tracking the path IDs though; this PR removes the unused overriding path IDs from the `GraphPath` data structure, which should save a bit of memory.