Errors while publishing coverage shouldn't block nightly builds
Right now, a failure to publish our codecov for a nightly build will fail
the nightly build entirely.  Note I am not saying "failure to meet codecov
requirements", but rather "when codecov cannot be transmitted, right now it
fails the build.".  Transient failures here are acceptable, and we get
notificiations when those are failed.