(fix) Rhai registration of setters even in non-settable context.
The upgrade of Rhai must have fixed the propagation https://rhai.rs/book/rust/getters-setters.html#chaining-updates so this was not creating an error at runtime because the call to split was causing a set on the supergraph headers resulting in a runtime error.

The fix is to remove all the fancy macro registration of handlers and just have simple code that registers appropriate handlers for each stage.