deps: remove axum-server
axum comes with a basic server implementation. axum-server allows much
more configuration, but we don't use any of it, so just `axum::serve`
will do!

`axum::serve` requires a `tokio::net::TcpListener` instead of a
`std::net::TcpListener` so some minor refactoring was required. The
public API of the snapshot server hasn't changed.

Practically this unblocks the `hyper` 1.8 upgrade, which has a breaking
change inside that's incompatible with `axum-server`.