Fix test_metrics() loop metric overlap
Test iterates through yaml files and generates an instrument for each config in the yaml.
Somehow after the upgrade this resulted in the metrics from the previous iteration of the loop
sticking around for the next iteration. The snapshot would then grow continually. This was fixed
by creating a new thread for each iteration, thus making a new meter provider per iteration