Fix empty `@join__directive(graphs: [], ...)` bug.
This PR proposes a fix for an apparent bug where the connectors
expansion code was creating `@join__directive` directives with empty
`graphs: []` arguments, accounting for (the only) composition errors in
at least 161/4530 supergraphs in our test corpus:
- Before (v2.12.1): 4283 successful, 247 failed
- After (this fix, based on current dev): 4444 successful, 86 failed

In an even more recent snapshot of this corpus, with 5741 supergraphs,
this fix reduced the number of failures from 707 to 236 failures, and
that only represents graphs where this empty `@join__directive(graphs: [])`
bug was the _only_ problem, suggesting this particular problem is
increasing in frequency and probably confusing new developers.

This code was originally related to license enforcement, but I believe
the plan of record is for connectors to be a free feature, so we can
probably do without any license enforcement here.

I'm certainly open to alternatives, including explanations of existing
behavior, but this seems like an accidental regression that is worth
fixing before (or as part of) the next connectors-related release.