fix(composition): Prune redundant join fields (#8744)
When applying `@join__field`, the `Merger` tries to limit applications to those where it adds information to the supergraph. When a field does not have `@join__field`, it is assumed to be present in all subgraphs. There are some cases where we add more applications than necessary. Instead of of changing the merge logic to be more selective, we can prune the schema after all merges are complete to remove this pattern. This specifically filters out `@join__field` applications when a field has `@join__field` for all subgraphs, and there is no other information (type, requires, etc) encoded in those `@join__field` applications.